+++ src/components/AdminTweetList.jsx
@@ -1,13 +1,11 @@
-import styled from 'styled-components';
 import { useEffect, useState } from 'react';
-import BeatLoader from 'react-spinners/BeatLoader';
 import { useNavigate } from 'react-router-dom';
-import Header from './Header';
-import { StyledListItem } from './TweetList';
+import BeatLoader from 'react-spinners/BeatLoader';
+import styled from 'styled-components';
+import { Header, StyledListItem, Alert } from './index';
 import { ReactComponent as Cross } from '../assets/Cross.svg';
 import { adminGetTweets, deleteTweet } from '../api/admin';
 import { countTimeDiff } from '../utilities';
-import Alert from './elements/Alert';
 
 // follow StyledListItem style by TweetList.jsx
 const StyledTweetContainer = styled(StyledListItem)`
