+++ src/pages/index.jsx
@@ -1,0 +1,8 @@
+export { default as AdminPage } from './AdminPage';
+export { default as HomePage } from './HomePage';
+export { default as LoginPage } from './LoginPage';
+export { default as ReplyPage } from './ReplyPage';
+export { default as SettingsPage } from './SettingsPage';
+export { default as SignupPage } from './SignupPage';
+export { default as TweetPage } from './TweetPage';
+export { default as UserPage } from './UserPage';
